+++ instrumentation/github.com/Shopify/sarama/option.go
@@ -28,8 +28,8 @@ const (
 )
 
 type config struct {
-	TraceProvider trace.Provider
-	Propagators   otelpropagation.Propagators
+	TracerProvider trace.Provider
+	Propagators    otelpropagation.Propagators
 
 	Tracer trace.Tracer
 }
@@ -37,14 +37,14 @@ type config struct {
 // newConfig returns a config with all Options set.
 func newConfig(opts ...Option) config {
 	cfg := config{
-		Propagators:   global.Propagators(),
-		TraceProvider: global.TraceProvider(),
+		Propagators:    global.Propagators(),
+		TracerProvider: global.TraceProvider(),
 	}
 	for _, opt := range opts {
 		opt(&cfg)
 	}
 
-	cfg.Tracer = cfg.TraceProvider.Tracer(defaultTracerName)
+	cfg.Tracer = cfg.TracerProvider.Tracer(defaultTracerName)
 
 	return cfg
 }
@@ -52,11 +52,11 @@ func newConfig(opts ...Option) config {
 // Option specifies instrumentation configuration options.
 type Option func(*config)
 
-// WithTraceProvider specifies a trace provider to use for creating a tracer for spans.
+// WithTracerProvider specifies a tracer provider to use for creating a tracer.
 // If none is specified, the global provider is used.
-func WithTraceProvider(provider trace.Provider) Option {
+func WithTracerProvider(provider trace.Provider) Option {
 	return func(cfg *config) {
-		cfg.TraceProvider = provider
+		cfg.TracerProvider = provider
 	}
 }