+++ instrumentation/github.com/astaxie/beego/config.go
@@ -21,56 +21,52 @@ import (
 	"go.opentelemetry.io/otel/api/trace"
 )
 
-// Config provides configuration for the beego OpenTelemetry
+// config provides configuration for the beego OpenTelemetry
 // middleware. Configuration is modified using the provided Options.
-type Config struct {
-	traceProvider trace.Provider
-	meterProvider metric.Provider
-	propagators   propagation.Propagators
-	filters       []Filter
-	formatter     SpanNameFormatter
+type config struct {
+	tracerProvider trace.Provider
+	meterProvider  metric.Provider
+	propagators    propagation.Propagators
+	filters        []Filter
+	formatter      SpanNameFormatter
 }
 
-// Option applies a configuration to the given Config.
+// Option applies a configuration to the given config.
 type Option interface {
-	Apply(*Config)
+	Apply(*config)
 }
 
 // OptionFunc is a function type that applies a particular
 // configuration to the beego middleware in question.
-type OptionFunc func(c *Config)
+type OptionFunc func(c *config)
 
-// Apply will apply the option to the Config, c.
-func (o OptionFunc) Apply(c *Config) {
+// Apply will apply the option to the config, c.
+func (o OptionFunc) Apply(c *config) {
 	o(c)
 }
 
 // ------------------------------------------ Options
 
-// WithTraceProvider sets the trace provider to be used by the middleware
-// to create a tracer for the spans.
-// Defaults to calling global.TraceProvider().
-// Tracer name is set to "go.opentelemetry.io/contrib/instrumentation/github.com/astaxie/beego".
-func WithTraceProvider(provider trace.Provider) OptionFunc {
-	return OptionFunc(func(c *Config) {
-		c.traceProvider = provider
+// WithTracerProvider specifies a tracer provider to use for creating a tracer.
+// If none is specified, the global provider is used.
+func WithTracerProvider(provider trace.Provider) Option {
+	return OptionFunc(func(cfg *config) {
+		cfg.tracerProvider = provider
 	})
 }
 
-// WithMeterProvider sets the meter provider to be used to create a meter
-// by the middleware.
-// Defaults to calling global.MeterProvider().
-// Meter name is set to "go.opentelemetry.io/contrib/instrumentation/github.com/astaxie/beego".
-func WithMeterProvider(provider metric.Provider) OptionFunc {
-	return OptionFunc(func(c *Config) {
-		c.meterProvider = provider
+// WithMeterProvider specifies a meter provider to use for creating a meter.
+// If none is specified, the global provider is used.
+func WithMeterProvider(provider metric.Provider) Option {
+	return OptionFunc(func(cfg *config) {
+		cfg.meterProvider = provider
 	})
 }
 
 // WithPropagators sets the propagators used in the middleware.
 // Defaults to global.Propagators().
 func WithPropagators(propagators propagation.Propagators) OptionFunc {
-	return OptionFunc(func(c *Config) {
+	return OptionFunc(func(c *config) {
 		c.propagators = propagators
 	})
 }
@@ -78,7 +74,7 @@ func WithPropagators(propagators propagation.Propagators) OptionFunc {
 // WithFilter adds the given filter for use in the middleware.
 // Defaults to no filters.
 func WithFilter(f Filter) OptionFunc {
-	return OptionFunc(func(c *Config) {
+	return OptionFunc(func(c *config) {
 		c.filters = append(c.filters, f)
 	})
 }
@@ -86,20 +82,20 @@ func WithFilter(f Filter) OptionFunc {
 // WithSpanNameFormatter sets the formatter to be used to format
 // span names. Defaults to the path template.
 func WithSpanNameFormatter(f SpanNameFormatter) OptionFunc {
-	return OptionFunc(func(c *Config) {
+	return OptionFunc(func(c *config) {
 		c.formatter = f
 	})
 }
 
 // ------------------------------------------ Private Functions
 
-func configure(options ...Option) *Config {
-	config := &Config{
-		traceProvider: global.TraceProvider(),
-		meterProvider: global.MeterProvider(),
-		propagators:   global.Propagators(),
-		filters:       []Filter{},
-		formatter:     defaultSpanNameFormatter,
+func configure(options ...Option) *config {
+	config := &config{
+		tracerProvider: global.TraceProvider(),
+		meterProvider:  global.MeterProvider(),
+		propagators:    global.Propagators(),
+		filters:        []Filter{},
+		formatter:      defaultSpanNameFormatter,
 	}
 	for _, option := range options {
 		option.Apply(config)