+++ instrumentation/go.mongodb.org/mongo-driver/config.go
@@ -23,32 +23,33 @@ const (
 	defaultTracerName = "go.opentelemetry.io/contrib/instrumentation/go.mongodb.org/mongo-driver"
 )
 
-// Config is used to configure the mongo tracer.
-type Config struct {
+// config is used to configure the mongo tracer.
+type config struct {
+	TracerProvider trace.Provider
+
 	Tracer trace.Tracer
 }
 
-// newConfig returns a Config with all Options set.
-func newConfig(opts ...Option) Config {
-	cfg := Config{}
+// newConfig returns a config with all Options set.
+func newConfig(opts ...Option) config {
+	cfg := config{
+		TracerProvider: global.TraceProvider(),
+	}
 	for _, opt := range opts {
 		opt(&cfg)
 	}
-	if cfg.Tracer == nil {
-		cfg.Tracer = global.Tracer(defaultTracerName)
-	}
+
+	cfg.Tracer = cfg.TracerProvider.Tracer(defaultTracerName)
 	return cfg
 }
 
 // Option specifies instrumentation configuration options.
-type Option func(*Config)
+type Option func(*config)
 
-// WithTracer specifies a tracer to use for creating spans. If none is
-// specified, a tracer named
-// "go.opentelemetry.io/contrib/instrumentation/go.mongodb.org/mongo-driver"
-// from the global provider is used.
-func WithTracer(tracer trace.Tracer) Option {
-	return func(cfg *Config) {
-		cfg.Tracer = tracer
+// WithTracerProvider specifies a tracer provider to use for creating a tracer.
+// If none is specified, the global provider is used.
+func WithTracerProvider(provider trace.Provider) Option {
+	return func(cfg *config) {
+		cfg.TracerProvider = provider
 	}
 }