+++ instrumentation/net/http/handler_test.go
@@ -43,8 +43,8 @@ func assertMetricLabels(t *testing.T, expectedLabels []label.KeyValue, measureme
 func TestHandlerBasics(t *testing.T) {
 	rr := httptest.NewRecorder()
 
-	tracer := mocktrace.Tracer{}
-	meterimpl, meter := mockmeter.NewMeter()
+	tracerProvider, tracer := mocktrace.NewProviderAndTracer(instrumentationName)
+	meterimpl, meterProvider := mockmeter.NewProvider()
 
 	operation := "test_handler"
 
@@ -57,8 +57,8 @@ func TestHandlerBasics(t *testing.T) {
 				t.Fatal(err)
 			}
 		}), operation,
-		WithTracer(&tracer),
-		WithMeter(meter),
+		WithTracerProvider(tracerProvider),
+		WithMeterProvider(meterProvider),
 	)
 
 	r, err := http.NewRequest(http.MethodGet, "http://localhost/", strings.NewReader("foo"))
@@ -102,7 +102,7 @@ func TestHandlerBasics(t *testing.T) {
 func TestHandlerNoWrite(t *testing.T) {
 	rr := httptest.NewRecorder()
 
-	tracer := mocktrace.Tracer{}
+	tracerProvider, tracer := mocktrace.NewProviderAndTracer(instrumentationName)
 
 	operation := "test_handler"
 	var span trace.Span
@@ -111,7 +111,7 @@ func TestHandlerNoWrite(t *testing.T) {
 		http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 			span = trace.SpanFromContext(r.Context())
 		}), operation,
-		WithTracer(&tracer),
+		WithTracerProvider(tracerProvider),
 	)
 
 	r, err := http.NewRequest(http.MethodGet, "http://localhost/", nil)
@@ -141,7 +141,7 @@ func TestHandlerNoWrite(t *testing.T) {
 func TestResponseWriterOptionalInterfaces(t *testing.T) {
 	rr := httptest.NewRecorder()
 
-	tracer := mocktrace.Tracer{}
+	tracerProvider, _ := mocktrace.NewProviderAndTracer(instrumentationName)
 
 	// ResponseRecorder implements the Flusher interface. Make sure the
 	// wrapped ResponseWriter passed to the handler still implements
@@ -155,7 +155,7 @@ func TestResponseWriterOptionalInterfaces(t *testing.T) {
 				t.Fatal(err)
 			}
 		}), "test_handler",
-		WithTracer(&tracer))
+		WithTracerProvider(tracerProvider))
 
 	r, err := http.NewRequest(http.MethodGet, "http://localhost/", nil)
 	if err != nil {